Changed how rankings are calculated for the profile page. It performs quicker, but, it's not as fast as I want and it doesn't rank according to the need.
The ranking approach that was in my head was more like an arcade highscore list. But the community was requesting just top. So, it doesn't matter if a person technically holds the 1st, 2nd, and 3rd places. They should only be attributed to the 1st place and the rest would still be up for grabs.